requirements.txt
django==4.2
django-filter>=2.4.0
djangorestframework>=3.11.0
gunicorn==20.1.0
psycopg2==2.8.5